Visual Studio Builds: Speed Up Release Builds
Use the multiprocessor compilation (/MP) option so that release build times
can be cut down quite a bit.  This will generate a brief warning for debug
builds as such builds use /Gm, but otherwise the build will proceed
normally albeit it would be slower.

Also use the /d2Zi+ flag for Visual Studio 2010 (and later) builds to log
more useful information in the .pdb files that are generated, to aid
debugging release builds when necessary.