MSVC Builds: Speed Up Release Builds
Use the multipropcessor compilation (/MP) option, which can help cut down
build times for release builds by quite a bit.  A warning will be emitted
for debug builds, due to the use of /Gm, but the build will otherwise
proceed normally.

Also use the /d2Zi+ compiler flag for MSVC 2010 (and later) builds, so that
more useful info would be logged to the .pdb files that are generated
during the build.